refactor(nexus_fs): extract InternalMixin + ContentMixin + MetadataMixin from monolithic NexusFS
Split the 5,122-line nexus_fs.py into domain-focused mixins:

- nexus_fs_internal.py (332 lines): shared helpers — context building,
  .readme overlay, path validation, post-write event dispatch
- nexus_fs_content.py (1,872 lines): content I/O — sys_read, sys_write,
  read_bulk, read_batch, stream, write_batch, edit, append, etc.
- nexus_fs_metadata.py (2,085 lines): metadata ops — sys_stat, sys_setattr,
  sys_unlink, sys_rename, sys_copy, sys_readdir, mkdir, stat_bulk, etc.
- nexus_fs.py (953 lines): lifecycle, IPC primitives, advisory locks,
  service forwarders, close/context manager

MRO: NexusFS(ContentMixin, MetadataMixin, InternalMixin, DispatchMixin, WatchMixin)

All method bodies preserved verbatim — pure structural refactor, no logic changes.
